external navigator was moved to common package
