+++ app/src/main/java/bes/max/playlistmaker/domain/mediateka/playlistdetails/SharePlaylistUseCase.kt
@@ -3,7 +3,7 @@ package bes.max.playlistmaker.domain.mediateka.playlistdetails
 import android.content.Context
 import bes.max.playlistmaker.R
 import bes.max.playlistmaker.domain.models.Playlist
-import bes.max.playlistmaker.domain.settings.ExternalNavigator
+import bes.max.playlistmaker.domain.common.ExternalNavigator
 
 class SharePlaylistUseCase(
     private val externalNavigator: ExternalNavigator,
